backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mne==1.5.1
numpy==1.24.3
scipy==1.11.3
matplotlib==3.7.2
pandas==2.0.3
scikit-learn==1.3.0
pyedflib==0.1.32
Pillow==10.0.1
celery==5.3.4
redis==5.0.1